third_party/blink/web_tests/http/tests/appcache/exceptions-expected.txt

CONSOLE WARNING: Application Cache API use is deprecated and will be removed in M82, around April 2020. See https://www.chromestatus.com/features/6192449487634432 for more details.
This tests that 'ApplicationCache' methods throw exceptions with reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cache.update() threw exception InvalidStateError: Failed to execute 'update' on 'ApplicationCache': there is no application cache to update..
PASS cache.swapCache() threw exception InvalidStateError: Failed to execute 'swapCache' on 'ApplicationCache': there is no newer application cache to swap to..
PASS successfullyParsed is true

TEST COMPLETE
